update mathjax namespace
